Create changeString.go
A program that takes three arguments: the name of a text file, and two
strings. This utility replaces every occurrence of the first string in the
file with the second string. For reasons of security, the final output will be
printed on the screen, which means that the original text file will remain intact.