Create kvStorage.go
 Program implements five fundamental tasks of a key-value store: adding, deleting, changing, looking up and printing data. The commands
for these four functions are "add", "del", "lookup" and "change" respectively.  The program will stop when you enter the "stop" word as input and will print the full contents of the key-value store when you enter the "print" command. Also there are used load() and save() functions to work with files. The "gob"
format is used to store the data.